+++ lib/clixon/clixon_validate.h
@@ -45,6 +45,7 @@
  */
 int xml_yang_validate_rpc(clicon_handle h, cxobj *xrpc, cxobj **xret);
 int xml_yang_validate_rpc_reply(clicon_handle h, cxobj *xrpc, cxobj **xret);
+int xml_yang_check_list_unique_minmax(cxobj *xt, cxobj **xret);
 int xml_yang_validate_add(clicon_handle h, cxobj *xt, cxobj **xret);
 int xml_yang_validate_list_key_only(cxobj *xt, cxobj **xret);
 int xml_yang_validate_all(clicon_handle h, cxobj *xt, cxobj **xret);