+++ test/test_datastore_format.sh
@@ -136,7 +136,7 @@ if [ $BE -ne 0 ]; then
 	err
     fi
     new "start backend -s init -f $cfg"
-    start_backend -s startup -f $cfg
+    start_backend -s init -f $cfg
 fi
 
 new "wait backend"